Fix active page highlighting in Header
The link to the currently active page wasn't correctly identified with the current path because of trailing slashes that are added when deployed. A helper function now normalizes the link and path with trailing slashes and returns whether the link is active. The link is also considered active if the current page is a child of the linked page.